+++ node_update.sh
@@ -34,7 +34,7 @@ if [[ $NODE_V != "v8."* ]]; then
       echo "Success! Network found."
 
       # make sure it keeps trying to fix if stopped mid progress
-      cp /home/pi/sisbot-server/sisbot/rc_fix.local /etc
+      cp /home/pi/sisbot-server/sisbot/rc_fix.local /etc/rc.local
 
       # install nodejs via apt-get and -yq yes and quit
       curl -sL https://deb.nodesource.com/setup_8.x | bash -